split forumCrud tests to make them easier to understand
