+++ backend/src/tests/Feature/ForumTests/ForumAnswerCreateTest.php
@@ -14,7 +14,7 @@ use App\Models\ContributorListProject;
 use App\Enums\ContributorStatusEnum;
 use Laravel\Sanctum\Sanctum;
 
-class ForumAnswerCrudTest extends TestCase
+class ForumAnswerCreateTest extends TestCase
 {
     use RefreshDatabase;
 
@@ -155,152 +155,5 @@ class ForumAnswerCrudTest extends TestCase
         );
 
         $response->assertStatus(422);
-    }
-
-    public function test_forum_answer_author_can_update(): void
-    {
-        $answer = ForumAnswer::factory()->create([
-            'forum_question_id' => $this->question->id,
-            'user_id' => $this->member->id,
-        ]);
-
-        Sanctum::actingAs($this->member);
-
-        $response = $this->putJson(
-            "/api/codeconnect/{$this->project->id}/forum/{$this->question->id}/answers/{$answer->id}",
-            ['answer' => 'Respuesta editada.']
-        );
-
-        $response->assertStatus(200)
-            ->assertJson([
-                'success' => true,
-                'message' => 'Answer updated successfully.',
-            ]);
-    }
-
-    public function test_forum_answer_owner_cannot_update(): void
-    {
-        $answer = ForumAnswer::factory()->create([
-            'forum_question_id' => $this->question->id,
-            'user_id' => $this->member->id,
-        ]);
-
-        Sanctum::actingAs($this->owner);
-
-        $response = $this->putJson(
-            "/api/codeconnect/{$this->project->id}/forum/{$this->question->id}/answers/{$answer->id}",
-            ['answer' => 'Editado por el owner.']
-        );
-
-        $response->assertStatus(403)
-            ->assertJson([
-                'success' => false,
-                'message' => 'You do not have permission to edit this answer.',
-            ]);
-    }
-
-    public function test_forum_answer_random_user_cannot_update(): void
-    {
-        $answer = ForumAnswer::factory()->create([
-            'forum_question_id' => $this->question->id,
-            'user_id' => $this->member->id,
-        ]);
-
-        $randomUser = User::factory()->create();
-        Sanctum::actingAs($randomUser);
-
-        $response = $this->putJson(
-            "/api/codeconnect/{$this->project->id}/forum/{$this->question->id}/answers/{$answer->id}",
-            ['answer' => 'Intento.']
-        );
-
-        $response->assertStatus(403)
-            ->assertJson([
-                'success' => false,
-            ]);
-    }
-
-    public function test_forum_answer_author_can_delete(): void
-    {
-        $answer = ForumAnswer::factory()->create([
-            'forum_question_id' => $this->question->id,
-            'user_id' => $this->member->id,
-        ]);
-
-        Sanctum::actingAs($this->member);
-
-        $response = $this->deleteJson(
-            "/api/codeconnect/{$this->project->id}/forum/{$this->question->id}/answers/{$answer->id}"
-        );
-
-        $response->assertStatus(200)
-            ->assertJson([
-                'success' => true,
-                'message' => 'Answer deleted successfully.',
-            ]);
-
-        $this->assertDatabaseMissing('forum_answers', [
-            'id' => $answer->id,
-        ]);
-    }
-
-    public function test_forum_answer_owner_can_delete(): void
-    {
-        $answer = ForumAnswer::factory()->create([
-            'forum_question_id' => $this->question->id,
-            'user_id' => $this->member->id,
-        ]);
-
-        Sanctum::actingAs($this->owner);
-
-        $response = $this->deleteJson(
-            "/api/codeconnect/{$this->project->id}/forum/{$this->question->id}/answers/{$answer->id}"
-        );
-
-        $response->assertStatus(200)
-            ->assertJson([
-                'success' => true,
-            ]);
-
-        $this->assertDatabaseMissing('forum_answers', [
-            'id' => $answer->id,
-        ]);
-    }
-
-    public function test_forum_answer_random_user_cannot_delete(): void
-    {
-        $answer = ForumAnswer::factory()->create([
-            'forum_question_id' => $this->question->id,
-            'user_id' => $this->member->id,
-        ]);
-
-        $randomUser = User::factory()->create();
-        Sanctum::actingAs($randomUser);
-
-        $response = $this->deleteJson(
-            "/api/codeconnect/{$this->project->id}/forum/{$this->question->id}/answers/{$answer->id}"
-        );
-
-        $response->assertStatus(403)
-            ->assertJson([
-                'success' => false,
-            ]);
-
-        $this->assertDatabaseHas('forum_answers', [
-            'id' => $answer->id,
-        ]);
-    }
-
-    public function test_forum_answer_unauthenticated_user_cannot_delete(): void
-    {
-        $answer = ForumAnswer::factory()->create([
-            'forum_question_id' => $this->question->id,
-        ]);
-
-        $response = $this->deleteJson(
-            "/api/codeconnect/{$this->project->id}/forum/{$this->question->id}/answers/{$answer->id}"
-        );
-
-        $response->assertStatus(401);
     }
 }
